:sparkles: Working invite user with success handling
